detect cycles. add pytest.
